techDebt(buRefactorTraining): Create high level validateTraining function in new file and move in mapping to run line validation
